Move controller classes
-  Create package for controller classes and move them there
