+++ src/main/java/org/phoebus/channelfinder/Application.java
@@ -43,7 +43,7 @@ import org.springframework.util.FileCopyUtils;
 @SpringBootApplication
 public class Application implements ApplicationRunner {
 
-  static final Logger logger = Logger.getLogger(Application.class.getName());
+  public static final Logger logger = Logger.getLogger(Application.class.getName());
 
   public static void main(String[] args) {
     // Set the java truststore used by channelfinder
