+++ src/main/java/org/phoebus/channelfinder/rest/controller/TagManager.java
@@ -1,4 +1,4 @@
-package org.phoebus.channelfinder;
+package org.phoebus.channelfinder.rest.controller;
 
 import static org.phoebus.channelfinder.CFResourceDescriptors.TAG_RESOURCE_URI;
 
@@ -20,6 +20,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
+import org.phoebus.channelfinder.TextUtil;
 import org.phoebus.channelfinder.entity.Channel;
 import org.phoebus.channelfinder.entity.Tag;
 import org.phoebus.channelfinder.respository.ChannelRepository;
