+++ src/test/java/org/phoebus/channelfinder/ChannelScrollIT.java
@@ -12,6 +12,7 @@ import org.phoebus.channelfinder.entity.Scroll;
 import org.phoebus.channelfinder.respository.ChannelRepository;
 import org.phoebus.channelfinder.respository.PropertyRepository;
 import org.phoebus.channelfinder.respository.TagRepository;
+import org.phoebus.channelfinder.rest.controller.ChannelScroll;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.test.context.TestPropertySource;
