+++ src/test/java/org/phoebus/channelfinder/docker/ChannelFinderScrollIT.java
@@ -25,6 +25,7 @@ import java.net.HttpURLConnection;
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Test;
 import org.phoebus.channelfinder.entity.Scroll;
+import org.phoebus.channelfinder.rest.controller.ChannelScroll;
 import org.testcontainers.containers.ComposeContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
@@ -34,7 +35,7 @@ import org.testcontainers.junit.jupiter.Testcontainers;
  * org.phoebus.channelfinder.CFResourceDescriptors#SCROLL_RESOURCE_URI}.
  *
  * @author Lars Johansson
- * @see org.phoebus.channelfinder.ChannelScroll
+ * @see ChannelScroll
  * @see org.phoebus.channelfinder.docker.ITUtil
  * @see org.phoebus.channelfinder.docker.ITUtilScroll
  */
