+++ src/test/java/org/phoebus/channelfinder/epics/EpicsRPCRequestIT.java
@@ -15,9 +15,9 @@ import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
-import org.phoebus.channelfinder.ChannelManager;
-import org.phoebus.channelfinder.PropertyManager;
-import org.phoebus.channelfinder.TagManager;
+import org.phoebus.channelfinder.rest.controller.ChannelManager;
+import org.phoebus.channelfinder.rest.controller.PropertyManager;
+import org.phoebus.channelfinder.rest.controller.TagManager;
 import org.phoebus.channelfinder.entity.Channel;
 import org.phoebus.channelfinder.entity.Property;
 import org.phoebus.channelfinder.entity.Tag;
