+++ src/test/java/org/phoebus/channelfinder/processors/ChannelProcessorManagerIT.java
@@ -10,7 +10,8 @@ import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.Mockito;
 import org.phoebus.channelfinder.CFResourceDescriptors;
-import org.phoebus.channelfinder.ChannelScroll;
+import org.phoebus.channelfinder.rest.controller.ChannelProcessorManager;
+import org.phoebus.channelfinder.rest.controller.ChannelScroll;
 import org.phoebus.channelfinder.entity.Scroll;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
