+++ src/test/resources/logging.properties
@@ -2,7 +2,7 @@ handlers= java.util.logging.ConsoleHandler
 
 .level = SEVERE
 
-org.phoebus.channelfinder.ChannelManager = SEVERE
+org.phoebus.channelfinder.rest.controller.ChannelManager = SEVERE
 o.p.channelfinder.ChannelManager = SEVERE
 org.phoebus.channelfinder.respository.ChannelRepository = SEVERE
 o.p.channelfinder.ChannelRepository = SEVERE
